refactored the debounce hook
