+++ src/actions/getCoinCarouselData.ts
@@ -8,7 +8,7 @@ import { CoinsListMarketData } from "@/utils/types/CoinsListMarketData";
 export const getCarouselSliderData = async () => {
   const response: CoinsListMarketData = await fetchData(
     "coins/markets?vs_currency=usd&per_page=100",
-    coins_market_data,
+    { localData: coins_market_data },
   );
 
   return parseCarouselSliderData(response);
