+++ src/actions/getSearchBarData.ts
@@ -1,7 +1,6 @@
 "use server";
 
 import { fetchData } from "@/utils/fetchData";
-// import { search_bar } from "@/utils/mockData/search_bar";
 import { SearchBarData } from "@/utils/types/SearchBarData";
 
 export const getSearchBarData = async (query: string) => {
