+++ src/actions/getStatsBarData.ts
@@ -6,6 +6,8 @@ import { parseStatsBarData } from "@/utils/parseStatsBarData";
 import { stats_bar } from "@/utils/mockData/status_bar";
 
 export const getStatsBarData = async () => {
-  const response: StatsBarData = await fetchData("global", stats_bar);
+  const response: StatsBarData = await fetchData("global", {
+    localData: stats_bar,
+  });
   return parseStatsBarData(response);
 };
