+++ src/components/PageHeader/NavBar/SearchBar.tsx
@@ -1,24 +1,12 @@
 "use client";
 
-import { useEffect, useState } from "react";
-import { useDebounce } from "@/hooks/useDebounce";
+import { useDebouncedInput } from "@/hooks/useDebouncedInput ";
 import { Input } from "@/components/ui/shadcn/input";
 import { SearchIcon } from "@/components/PageHeader/NavBar/icons";
-import { getSearchBarData } from "@/actions/getSearchBarData";
-import { SearchBarData } from "@/utils/types/SearchBarData";
 
 const SearchBar = () => {
-  const [data, setData] = useState<SearchBarData>(null);
-  const { value, debouncedValue, handleChange } = useDebounce(250);
-
-  useEffect(() => {
-    if (!debouncedValue) {
-      setData(null);
-      return;
-    }
-
-    getSearchBarData(debouncedValue).then((response) => setData(response));
-  }, [debouncedValue]);
+  const { data, value, handleChange, clearSearchResults } =
+    useDebouncedInput(250);
 
   return (
     <div className="relative w-80 [&>svg]:absolute [&>svg]:left-0 [&>svg]:top-2.5 [&>svg]:ml-3">
@@ -26,7 +14,7 @@ const SearchBar = () => {
         value={value}
         placeholder="Search..."
         onChange={handleChange}
-        onBlur={() => setData(null)}
+        onBlur={clearSearchResults}
         className="w-full border-none bg-[var(--clr-nav-foreground)] pl-10 pr-4 placeholder:text-inherit hover:bg-[var(--clr-hover)] focus-visible:ring-0"
       />
       <ul className="scrollbar absolute top-9 z-50 -mt-1 max-h-96 w-full overflow-y-auto bg-[var(--clr-nav-foreground)]">
